Add some preliminary code for MIDI 2.0 extension support.
To get MIDI 2.0 support working, we need some host that can send UMPs on
the audio plugin buffers (which is most likely aaphostsample for our case),
as well as some audio plugin that actually receives MIDI 2.0 messages
(which is most likely aap-ayumi for our case.

An experimental host sample would send some pitch 7.9 note messages to the
plugin and the plugin would process the message and generate microtonal output.