v0.1.13: aap-ayumi works as a MidiDeviceService, and is ready in MIDI 2.0 mode.
It makes use of the new MidiDeviceService as a library in aap-ayumi.

atsushieno/kmmk can send MIDI 2.0 UMPs over Android MIDI API for dogfooding.

The MidiDeviceService implementation is still not ready for MIDI 1.0, but
any aap-lv2 instrument can be MIDI 2.0 ready as long as it works with the
downconverted MIDI 1.0 messages.