aap-import-lv2-metadata: toggle is not necessarily a bool value. Parse as float.
This fixes the issue that MDA EPiano generated only one note. It was
incorrectly setting `Polyphonic` port property as 0.0 = disabled.

It was caused by the port default value interpreter, which was parsing
the value as RDF boolean for port that has `toggled` property.

The LV2 Core specification says about `toggled`:

> Indicates that the data item should be considered a boolean toggle.
> Data less than or equal to zero should be considered "off" or "false",
> and data above zero should be considered "on" or "true".

I thought `toggled` is a boolean property, but `lilv_node_as_bool()` treats
the value if and only if it is "true". Since this function does not
particularly target `toggled`, it seems to target "indeed boolean" values,
not the type of LV2 port default property (which is just "Datatype").

Therefore, the value should be interpreted as a number (float), and then
it should be interpreted as the documented way ( true if x>0, otherwise false).