Rewrite Atom_Sequence processing to handle ResizePort minimumSize appropriately.
In earlier code, Atom output ports were mostly disregarded and left untouched,
and only MIDI input atom ports were paid attention.

But the latest sfizz has automation ports that contain values for MIDI
control changes, which go beyond the default buffer size (which used to be
like 4096 * 4 bytes). For sfizz, the output Atom buffer size, as well as
the atom sequence size is not expected as it specifies the automation port
size with ResizePort::minimumSize (as 524288).

I'm not sure if a plugin expecting Atom sequence size information on the
output port to indicate the host buffer capacity is appropriate, but
we could fill it so far.