fix state and preset restore feature that got broken after "V2" protocol changes.
We used to simply map LV2 ports to AAP ports. Now we don't, and those
control port buffer went to simple array of float. So IF the restored
port is float control data, then it should work. IF it is not but still
assigned to Atom sequence (like Patch:Message), there is still some hope
that it may work (unless some conflicting message is passed there).
For other case, it should be said not supported...

seealso: https://github.com/atsushieno/aap-lv2/issues/7

Also, preset restoring had an issue that it was not passing the right URI.