fix build; material3 reference is left awkward (as it is resolved via deps)
androidaudioplugin-ui-compose-app depends on material3 but libs.versions.toml
does not specify the version. It seems wrong, but we cannot simply reference
the version because it is resolved via BOM.

In aap-core, it is resolved via compose-audio-controls so the lack of the
version number is not problematic it seems.
aap-ayumi was referencing it directly, without compose-audio-controls, thus
it resulted in unresolved reference. We have no direct reference to it, so
simply remove it from the reference list.